+++ eCommerce/Inventory.h
@@ -9,7 +9,7 @@ class Inventory {
     private:
         float revenue;
     public:
-        Inventory(const map<Item*, int>& stock={}, const int& revenue=0);
+        Inventory(const map<Item*, int>& stock={}, const float& revenue=0);
         bool add(const string& name, const int& quantity=1, const float& price=-1);
         bool remove(const string& name, const int& quantity=1);
         float getRevenue();
